fix: 라우팅 시 스타일이 적용되지 않는 오류 수정
