--- src/libs/useServerInsertedRegistry.tsx
@@ -1,17 +1,0 @@
-'use client';
-
-import { useServerInsertedHTML } from 'next/navigation';
-import { useStyledComponentsRegistry } from './styledComponent';
-
-export default function RootStyleRegistry({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  const [StyledComponentsRegistry, styledComponentsFlushEffect] =
-    useStyledComponentsRegistry();
-
-  useServerInsertedHTML(() => <>{styledComponentsFlushEffect()}</>);
-
-  return <StyledComponentsRegistry>{children}</StyledComponentsRegistry>;
-}
